backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
slowapi==0.1.9
google-generativeai==0.5.0
razorpay==1.4.1
websockets==12.0
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.2.0
